Renamed App.config to config
